Almost working, but optim constraints don't show up yet
